+++ src/main/java/me/pajic/simple_smithing_overhaul/mixin/ItemStackMixin.java
@@ -2,16 +2,27 @@ package me.pajic.simple_smithing_overhaul.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import me.pajic.simple_smithing_overhaul.Main;
 import me.pajic.simple_smithing_overhaul.util.ModDataComponents;
 import me.pajic.simple_smithing_overhaul.util.ModUtil;
 import net.minecraft.ChatFormatting;
+import net.minecraft.core.Holder;
 import net.minecraft.core.component.DataComponentHolder;
 import net.minecraft.core.component.DataComponentType;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.EquipmentSlotGroup;
+import net.minecraft.world.entity.ai.attributes.Attribute;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.component.ItemAttributeModifiers;
+import org.apache.commons.lang3.function.TriConsumer;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -19,6 +30,8 @@ import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import java.util.function.BiConsumer;
 
 @Mixin(ItemStack.class)
 public abstract class ItemStackMixin implements DataComponentHolder {
@@ -37,6 +50,12 @@ public abstract class ItemStackMixin implements DataComponentHolder {
 
     @Shadow
     public abstract boolean isDamageableItem();
+
+    @Shadow
+    public abstract boolean is(Item item);
+
+    @Shadow
+    public abstract int getDamageValue();
 
     @Unique
     private final ItemStack thisStack = (ItemStack) (Object) this;
@@ -71,12 +90,32 @@ public abstract class ItemStackMixin implements DataComponentHolder {
         return !ModUtil.isBroken(instance);
     }
 
+    @WrapMethod(method = "forEachModifier(Lnet/minecraft/world/entity/EquipmentSlot;Ljava/util/function/BiConsumer;)V")
+    private void noAttributesIfBroken(EquipmentSlot equipmentSLot, BiConsumer<Holder<Attribute>, AttributeModifier> action, Operation<Void> original) {
+        if (!ModUtil.isBroken(thisStack)) original.call(equipmentSLot, action);
+    }
+
+    @WrapMethod(
+            //? if < 1.21.8
+            method = "forEachModifier(Lnet/minecraft/world/entity/EquipmentSlotGroup;Ljava/util/function/BiConsumer;)V"
+            //? if >= 1.21.8
+            /*method = "forEachModifier(Lnet/minecraft/world/entity/EquipmentSlotGroup;Lorg/apache/commons/lang3/function/TriConsumer;)V"*/
+    )
+    private void noAttributesIfBroken(
+            //? if < 1.21.8
+            EquipmentSlotGroup slotGroup, BiConsumer<Holder<Attribute>, AttributeModifier> action, Operation<Void> original
+            //? if >= 1.21.8
+            /*EquipmentSlotGroup slotGroup, TriConsumer<Holder<Attribute>, AttributeModifier, ItemAttributeModifiers.Display> action, Operation<Void> original*/
+    ) {
+        if (!ModUtil.isBroken(thisStack)) original.call(slotGroup, action);
+    }
+
     @ModifyReturnValue(
             method = "getHoverName",
             at = @At("RETURN")
     )
     private Component changeNameIfBroken(Component original) {
-        if (ModUtil.isBroken(thisStack)) {
+        if (ModUtil.isBroken(thisStack) || (is(Items.ELYTRA) && getMaxDamage() - getDamageValue() == 1)) {
             MutableComponent name = Component.translatable("item.simple_smithing_overhaul.broken");
             name.append(original);
             name.withStyle(ChatFormatting.RED);